tests: config_format: more compare methods to test improvements
Strictly OK/FAIL when the before-after don't match is less useful
when we are making improvements that introduce changes on purpose.
Change labels to MATCH/NO MATCH, and also add -c which shows the
wc of output lines - much easier to confirm when updates cause
more tests to successfully match.

Signed-off-by: Hank Leininger <hlein@korelogic.com>